+++ src/parser/features/DDBAttackAction.js
@@ -35,7 +35,7 @@ export default class DDBAttackAction extends DDBAction {
       this._generateProperties();
       this._generateSystemType(this.type);
       this._generateSystemSubType();
-      if (!this.enricher.documentStub?.stopDefaultActivity)
+      if (!this.enricher.stopDefaultActivity)
         this._generateActivity();
       await this.enricher.addAdditionalActivities(this);